añadimos registro
