+++ .gitignore
@@ -1,0 +1,2 @@
+/ValidarCampos.class
+/registroUsuario.class
